pyproject.toml
[project]
name = "alpaca-paper-day-bot"
version = "0.1.0"
description = "Local-only Alpaca paper day-trading bot (1-week mock run)."
requires-python = ">=3.10"
dependencies = [
  "alpaca-py>=0.42.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "pytz>=2024.1",
  "pandas-ta>=0.3.14b0",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "rich>=13.7",
]

[project.scripts]
alpaca-day-bot = "alpaca_day_bot.main:cli"
alpaca-watch-trades = "alpaca_day_bot.watch_trades:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

